moves validator code into prompting package
