+++ prompting/__init__.py
@@ -36,6 +36,7 @@ from . import forward
 from . import agent
 from . import conversation
 from . import dendrite
+from . import validator
 
 from .llms import hf
